feat(cloud-cost-pricing): extract AwsPricingApiAdapter.warmUp into the shared package
cloudrift-iac-detector needs live regional pricing for its zombie-cost
estimate, but only the generic warmUp()/fetchPrice() prefetch — not the
~13 per-instance-type on-demand methods core's own underutilized scanners
use. New minimal AwsPricingApiAdapter (warmUp/fetchPrice/dispose only) in
libs/shared/cloud-cost-pricing, exported from @ellevas/cloud-cost-pricing.
Core's own adapter now composes an instance of it for warmUp, sharing one
PricingClient, instead of duplicating the prefetch/fetch logic; its 13
on-demand methods are unchanged. Also fixes a latent shared-aws-infra-utils
moduleNameMapper gap in 4 jest configs, surfaced now that cloud-cost-pricing
depends on it transitively.
