+++ libs/dead-resources/domain/jest.config.cjs
@@ -18,6 +18,8 @@ module.exports = {
   moduleNameMapper: {
     '^shared-kernel$': '<rootDir>/../../shared/kernel/src/index.ts',
     '^shared-kernel/(.*)$': '<rootDir>/../../shared/kernel/src/$1',
+    '^shared-aws-infra-utils$': '<rootDir>/../../shared/aws-infra-utils/src/index.ts',
+    '^shared-aws-infra-utils/(.*)$': '<rootDir>/../../shared/aws-infra-utils/src/$1',
     '^cloud-cost-domain$': '<rootDir>/../../cloud-cost/domain/src/index.ts',
     '^cloud-cost-domain/(.*)$': '<rootDir>/../../cloud-cost/domain/src/$1',
     '^cloud-cost-pricing$': '<rootDir>/../../shared/cloud-cost-pricing/src/index.ts',
